WIP refactoring for .NET 7 JavascriptServices interop
