+++ SerratedJQLibrary/SerratedJQ/EmbeddedFiles.resx
@@ -119,7 +119,10 @@
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
   <data name="DeprecatedHttpClient" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>Resources\HttpClient.txt;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
+    <value>Resources\HttpClient.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
+  </data>
+  <data name="JQueryProxy" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>EmbeddedJavascript\JQueryProxy.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
   <data name="ObserveRemovedJs" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>EmbeddedJavascript\ObserveRemoved.js;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
