+++ SerratedJQLibrary/SerratedJQ/JQuerySelectorBuilder.cs
@@ -1,0 +1,98 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SerratedSharp.SerratedJQ
+{
+    
+    // Implement a class that exposes a fluent API for building jQuery selectors based on https://api.jquery.com/category/selectors/
+    //public class JQuerySelectorBuilder
+    //{
+    //    public JQuerySelectorBuilder()
+    //    {
+
+    //    }
+
+    //    public JQuerySelectorBuilder Id(string id)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Class(string className)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName, string attributeValue)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName, string attributeValue, string attributeOperator)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName, string attributeValue, string attributeOperator, string attributeModifier)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName, string attributeValue, string attributeOperator, string attributeModifier, string attributeValueModifier)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Attribute(string attributeName, string attributeValue, string attributeOperator, string attributeModifier, string attributeValueModifier, string attributeValueModifierValue)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelectorBuilder Element(string elementName)
+    //    {
+    //        return this;
+    //    }
+
+    //    public JQuerySelector GetSelector()
+    //    {
+    //        return new JQuerySelector();
+    //    }
+
+    //}
+
+    //public class JQuerySelector
+    //{
+    //    public JQuerySelector()
+    //    {
+            
+    //    }
+
+    //    string cachedSelector = null;
+
+    //    public string Render()
+    //    {
+    //        if(cachedSelector == null)
+    //        {
+    //            cachedSelector = RenderSelector();
+    //        }
+            
+    //        return cachedSelector;
+            
+    //    }
+
+    //    protected virtual string RenderSelector()
+    //    {
+    //        return "";
+    //    }
+
+
+    //}
+
+}
