+++ SerratedJQLibrary/Tests.Wasm/Properties/launchSettings.json
@@ -2,7 +2,12 @@
   "profiles": {
     "Tests.Wasm": {
       "commandName": "Project",
-      "launchBrowser": true
+      "launchBrowser": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "inspectUri": "{wsProtocol}://{url.hostname}:{url.port}/_framework/debug/ws-proxy?browser={browserInspectUri}",
+      "applicationUrl": "https://localhost:56546;http://localhost:56547"
     }
   }
 }
